requirements.txt
Django~=4.0.3
matplotlib~=3.5.1
django-crispy-forms
django_pdb
django-heroku
djangorestframework
django-rest-swagger
requests
gunicorn
asgiref==3.4.1
spacy~=3.4.1
openai~=0.22.1
python-dotenv
chronological
stripe
drf_yasg